requirements.txt
# Semantic Translation MVP - The Semantic Bridge
# =============================================
#
# IMPORTANT NOTES:
# - penman is REQUIRED for AMR graph serialization (don't skip this!)
# - amrlib requires separate model download (see README)
# - If amrlib model is missing, system falls back to LLM-based parsing
#
# Installation: pip install -r requirements.txt
# =============================================

# Core AMR parsing
amrlib>=0.8.0
penman>=1.2.2          # Critical: AMR graph serialization
smatch>=1.0.4          # AMR graph comparison metric
unidecode>=1.3.0       # Required by amrlib's parser

# LLM Integration
openai>=1.0.0
anthropic>=0.18.0
google-genai>=1.0.0        # Google Gemini API

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.26.0

# For AMR visualization (optional)
networkx>=3.0
matplotlib>=3.8.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0

